refactor: 채팅방 입장 API 관련 도메인 chat->product로 이동
