+++ src/main/java/org/mwt/market/domain/chat/service/ChatService.java
@@ -39,40 +39,6 @@ public class ChatService {
         this.chatContentRepository = chatContentRepository;
     }
 
-    @Transactional
-    public ChatRoomDto joinChatRoom(UserPrincipal userPrincipal, Long productId) {
-        Long userId = userPrincipal.getId();
-
-        Optional<ChatRoom> optChatRoom = chatRoomRepository
-            .findByBuyer_UserIdAndProduct_ProductId(userId,
-                productId);
-
-        ChatRoom chatRoom = optChatRoom.orElseGet(() -> {
-            User buyer = userRepository.findById(userId)
-                .orElseThrow(NoSuchUserException::new);
-            Product product = productRepository.findById(productId)
-                .orElseThrow(NoSuchProductException::new);
-            return chatRoomRepository.save(ChatRoom.createChatRoom(buyer, product));
-        });
-
-        ChatContent lastChatContent = chatContentRepository
-                .findFirstByChatRoomIdOrderByCreateAtDesc(chatRoom.getChatRoomId());
-
-        String lastMessage = "";
-        if (lastChatContent != null) {
-            lastMessage = lastChatContent.getContent();
-        }
-
-        return ChatRoomDto.builder()
-            .chatRoomId(chatRoom.getChatRoomId())
-            .buyerId(chatRoom.getBuyer().getUserId())
-            .nickName(chatRoom.getBuyer().getNickname())
-            .buyerProfileImg(chatRoom.getBuyer().getProfileImageUrl())
-            .lastMessage(lastMessage)
-            .lastCreatedAt(chatRoom.getCreatedAt())
-            .build();
-    }
-
     public MessageResponse saveMessage(String roomId, MessageRequest messageRequest) {
 
         ChatContent chatContent = ChatContent.builder()
